+++ lib/l10n/arb/app_uk.arb
@@ -786,6 +786,16 @@
     "examples": "Текст для кнопки на сторінці Discount.",
     "description": "Кнопка використовується для відриття AdvancedFilter або відкриття діалогу з ним."
   },
+  "filter": "Фільтр",
+  "@filter": {
+    "examples": "Текст для кнопки на сторінці Discount.",
+    "description": "Кнопка використовується для відриття діалогу фільтрації."
+  },
+  "sort": "Сортувати",
+  "@sort": {
+    "examples": "Текст для кнопки на сторінці Discount.",
+    "description": "Кнопка використовується для відриття меню фільтрації."
+  },
   "specialOffers": "Спеціальні пропозиції",
   "@specialOffers": {
     "examples": "Текст частини заголовку на сторінці Discount."
