+++ lib/shared/widget/cancel_chip_widget.dart
@@ -56,7 +56,7 @@ class CancelChipWidget extends StatelessWidget {
         labelText,
         textAlign: TextAlign.center,
         maxLines: 1,
-        overflow: TextOverflow.ellipsis,
+        // overflow: TextOverflow.ellipsis,
         style: textStyle ??
             (isDesk
                 ? AppTextStyle.materialThemeBodyLarge
